backend/requirements.txt
Flask
Flask-CORS
python-dotenv
google-cloud-storage
google-generativeai
opencv-python-headless